Replace 'static_cast<size_t>(10)' with '10U'

+++ map/map_tf_generator/test/test_uniform_random.cpp
@@ -36,7 +36,7 @@ TEST(UniformRandom, UniformRandom)
 
     for (int i = 0; i < 50; i++) {
       const std::vector<size_t> random = UniformRandom(4, 10);
-      ASSERT_EQ(random.size(), static_cast<size_t>(10));
+      ASSERT_EQ(random.size(), 10U);
       ASSERT_THAT(random, Each(AllOf(Ge(min_inclusive), Lt(max_exclusive))));  // in range [0, 4)
     }
   }
